M4: solution - refine domain models and separate API DTOs

+++ src/TheButton.Api/Features/V2/Counter/CounterResponse.cs
@@ -1,3 +1,3 @@
 namespace TheButton.Api.Features.V2.Counter;
 
-public record CounterResponse(int Value);
+public record CounterResponse(long Value);
